// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/product/models/product.model.ts
export class CategoryEntityModel {
  public categoryId: string;
  public categoryName: string;
  public categoryCode: string;
  public categoryTypeId: string;
  public categoryTypeName: string;
  public createdById: string;
  public createdDate: Date;
  public updatedDate: Date;
  public updatedById: string;
  public active: boolean;
  public isEdit: boolean;
  public isDefault: boolean;
  public categoryTypeCode: string;
  public countCategoryById: number;
  public sortOrder: number;
  public isDefauld: boolean;
  public tenantId: string;
}
// ...
export class EmployeeEntityModel {
  employeeId: string;
  mission: string;
  employeeName: string;
  employeeCode: string;
  OrganizationId: string;
  PositionId: string;
  EmployeeCodeName: string;
  roleId : string | null;
  isDisable : boolean | null;
  value : string | null;
  chucVuId: number;
}
// ...
export class TrangThaiGeneral {
  value: number;
  valueText: string;
  name: string;
  objectRole : TrangThaiGeneral;
  listEmployeeEntityModel : EmployeeEntityModel[];
  isEdit : boolean | null;
  categoryId : string | null;
}